Separated idl to c code generator off into own library
The C code generator is now no longer "included" in idlc, but
inside its own library just like all other code generators
This required the generation of extra export headers in addition
to the restructuring of the code
Some of the non-C generators will need to be modified slightly
and include the C generator library, as this now includes the
type_meta information generation functions
