+++ src/tools/idlc/include/libidlc/idl_defs.h
@@ -1,4 +1,4 @@
-// Copyright(c) 2020 to 2022 ZettaScale Technology and others
+// Copyright(c) 2023 ZettaScale Technology and others
 //
 // This program and the accompanying materials are made available under the
 // terms of the Eclipse Public License v. 2.0 which is available at
@@ -8,25 +8,38 @@
 //
 // SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
 
-#ifndef IDLC_GENERATOR_H
-#define IDLC_GENERATOR_H
+#ifndef IDL_DEFS_H
+#define IDL_DEFS_H
 
+#include <stdbool.h>
 #include <stdint.h>
-
 #include "idl/processor.h"
-#include "idl/tree.h"
-#include "idlc/options.h"
 
 #if defined (__cplusplus)
 extern "C" {
 #endif
 
-#define IDLC_GENERATOR_OPTIONS generator_options
-#define IDLC_GENERATOR_ANNOTATIONS generator_annotations
-#define IDLC_GENERATE generate
+typedef struct idlc_option idlc_option_t;
+struct generator;
 
-typedef struct idlc_generator_config idlc_generator_config_t;
-struct idlc_generator_config  {
+struct idlc_option {
+  enum {
+    IDLC_FLAG, /**< flag-only, i.e. (sub)option without argument */
+    IDLC_STRING,
+    IDLC_FUNCTION,
+  } type;
+  union {
+    int *flag;
+    const char **string;
+    int (*function)(const idlc_option_t *, const char *);
+  } store;
+  char option; /**< option, i.e. "o" in "-o". "-h" is reserved */
+  char *suboption; /**< name of suboption, i.e. "mount" in "-o mount" */
+  char *argument;
+  char *help;
+};
+
+typedef struct {
   char *output_dir; /* path to write completed files */
   char* base_dir; /* Path to start reconstruction of dir structure */
 
@@ -39,7 +52,7 @@ struct idlc_generator_config  {
   Passing a pointer to a generator function is a reasonable way of avoiding the layering problems
   this introduces. May be a null pointer */
   idl_retcode_t (*generate_typeinfo_typemap) (const idl_pstate_t *pstate, const idl_node_t *node, idl_typeinfo_typemap_t *result);
-};
+} idlc_generator_config_t;
 
 typedef const idlc_option_t **(*idlc_generator_options_t)(void);
 typedef const idl_builtin_annotation_t **(*idlc_generator_annotations_t)(void);
@@ -49,4 +62,4 @@ typedef int(*idlc_generate_t)(const idl_pstate_t *, const idlc_generator_config_
 }
 #endif
 
-#endif /* IDLC_GENERATOR_H */
+#endif /* IDL_DEFS_H */
