+++ src/tools/idlc/src/descriptor.c
@@ -27,7 +27,7 @@
 #include "generator.h"
 #include "descriptor.h"
 #include "hashid.h"
-#include "descriptor_type_meta.h"
+#include "libidlc/descriptor_type_meta.h"
 #include "dds/ddsc/dds_opcodes.h"
 
 #define TYPE (16)
@@ -2516,8 +2516,6 @@ ctype_fini(struct constructed_type *ctype)
   if (ctype->instructions.table)
    idl_free (ctype->instructions.table);
 }
-
-idl_retcode_t generate_descriptor(const idl_pstate_t *pstate, struct generator *generator, const idl_node_t *node);
 
 void
 descriptor_fini(struct descriptor *descriptor)
