+++ src/tools/idlc/src/descriptor_type_meta.c
@@ -33,8 +33,8 @@
 #include "idl/stream.h"
 #include "idl/string.h"
 #include "idl/misc.h"
-#include "generator.h"
 #include "descriptor_type_meta.h"
+#include "generator.h"
 
 static struct dds_cdrstream_allocator idlc_cdrstream_default_allocator = { idl_malloc, idl_realloc, idl_free };
