+++ src/tools/idlc/src/generator.h
@@ -14,7 +14,7 @@
 #include <stdio.h>
 
 #include "idl/processor.h"
-#include "idlc/generator.h"
+#include "libidlc/generator.h"
 
 #include <stdlib.h>
 #include <string.h>
@@ -30,28 +30,10 @@ struct generator {
     char *path;
   } source;
   struct {
-    struct idlc_generator_config c;
+    idlc_generator_config_t c;
     char *export_macro;
     bool generate_cdrstream_desc;
   } config;
 };
-
-int print_type(char *str, size_t len, const void *ptr, void *user_data);
-int print_scoped_name(char *str, size_t len, const void *ptr, void *user_data);
-
-#if _WIN32
-__declspec(dllexport)
-#endif
-const idlc_option_t** idlc_generator_options(void);
-
-#if _WIN32
-__declspec(dllexport)
-#endif
-idl_retcode_t idlc_generate(const idl_pstate_t *pstate, const idlc_generator_config_t *config);
-
-#if _WIN32
-__declspec(dllexport)
-#endif
-idl_retcode_t generate_nosetup(const idl_pstate_t *pstate, struct generator *generator);
 
 #endif /* GENERATOR_H */
