+++ src/tools/idlc/src/types.h
@@ -1,0 +1,18 @@
+// Copyright(c) 2023 ZettaScale Technology and others
+//
+// This program and the accompanying materials are made available under the
+// terms of the Eclipse Public License v. 2.0 which is available at
+// http://www.eclipse.org/legal/epl-2.0, or the Eclipse Distribution License
+// v. 1.0 which is available at
+// http://www.eclipse.org/org/documents/edl-v10.php.
+//
+// SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
+
+#ifndef TYPES_H
+#define TYPES_H
+
+#include "libidlc/idl_defs.h"
+
+idl_retcode_t generate_types(const idl_pstate_t *pstate, struct generator *generator);
+
+#endif /* TYPES_H */
